Supplier tooling + docs checkpoint
