+++ docusaurus/docs/operate/faq/2_validator_faq.md
