+++ docusaurus/docs/operate/faq/3_supplier_faq.md
@@ -1,0 +1,18 @@
+---
+sidebar_position: 3
+title: Supplier FAQ
+---
+
+Get diagrams from here:
+
+https://medium.com/decentralized-infrastructure/exploring-paths-capabilities-with-pocket-network-s-shannon-upgrade-and-deepseek-part-one-2a3ee6032dfb
+
+https://medium.com/decentralized-infrastructure/exploring-paths-capabilities-with-pocket-network-s-shannon-upgrade-and-deepseek-part-two-a5e38766fd71
+
+## What is the different between a RelayMiner & Supplier
+
+TODO: Add onchain / offchain diagram
+
+## What happens if you go below the Min stake?
+
+## What is the maximum stake?
